Sharing code between Pusher and PusherAsync
